requirements.txt
streamlit
openai
pypdf
pdf2image
pillow
colorthief
requests
python-dotenv
